python-llm/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
langchain==0.3.14
langchain-core>=0.3,<0.4
langchain-openai>=0.2,<0.3
langchain-community>=0.3,<0.4
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
